+++ frontend/src/features/material-details/components/DemandCategoryModal.tsx
@@ -99,6 +99,18 @@ const createDemandColumns = (handleDelete?: (row: Demand) => void) => {
                 );
             },
         },
+        {
+            field: 'lastUpdatedOnDateTime',
+            headerName: 'Updated',
+            headerAlign: 'center',
+            flex: 1.5,
+            renderCell: (data: { row: Demand }) => (
+                <Stack display="flex" textAlign="center" alignItems="center" justifyContent="center" width="100%" height="100%">
+                    <Box>{new Date(data.row.lastUpdatedOnDateTime).toLocaleDateString('en-GB')}</Box>
+                    <Box>{new Date(data.row.lastUpdatedOnDateTime).toLocaleTimeString('en-GB')}</Box>
+                </Stack>
+            ),
+        },
     ] as const;
     if (handleDelete) {
         return [