+++ frontend/src/features/material-details/components/PlannedProductionModal.tsx
@@ -87,6 +87,18 @@ const createProductionColumns = (handleDelete?: (row: Production) => void) => {
                 </Box>
             ),
         },
+        {
+            field: 'lastUpdatedOnDateTime',
+            headerName: 'Updated',
+            headerAlign: 'center',
+            flex: 1.5,
+            renderCell: (data: { row: Production }) => (
+                <Stack display="flex" textAlign="center" alignItems="center" justifyContent="center" width="100%" height="100%">
+                    <Box>{new Date(data.row.lastUpdatedOnDateTime).toLocaleDateString('en-GB')}</Box>
+                    <Box>{new Date(data.row.lastUpdatedOnDateTime).toLocaleTimeString('en-GB')}</Box>
+                </Stack>
+            ),
+        },
     ] as const;
     if (handleDelete) {
         return [