+++ CHANGELOG.md
@@ -20,6 +20,10 @@ The **need for configuration updates** is **marked bold**.
 * Use the frontend's existing keycloak authentication to authorize backend calls ([#896](https://github.com/eclipse-tractusx/puris/pull/896)) (**moved IDP configuration and IDP no longer optional**)
 * Rework validation logic to return detailed error messages ([#907](https://github.com/eclipse-tractusx/puris/pull/907))
 
+Fixes
+
+* Added lastUpdatedOnDateTime to local representations of Demand, Production and Delivery ([#908](https://github.com/eclipse-tractusx/puris/pull/908))
+
 ## [v3.1.0](https://github.com/eclipse-tractusx/puris/releases/tag/3.1.0)
 
 The following Changelog lists the changes. Please refer to the [documentation](docs/README.md) for configuration needs and understanding the concept changes.